v1111

+++ util/osUtil/fileDelete.go
@@ -9,6 +9,9 @@ import (
 
 func DeleteFile(TotalUrl string) {
 	pos := strings.Index(TotalUrl, "/uploadPic")
+	if pos == -1 {
+		return
+	}
 	finalPicUrl := config.PicSaverPath + TotalUrl[pos+10:]
 	err := os.Remove(finalPicUrl)
 	if err != nil {
