+++ util/saveUtil/delete.go
@@ -10,6 +10,9 @@ import (
 // DeletePicSaver 删除数据库存储在本地的图片
 func DeletePicSaver(picUrl string) (err error) {
 	pos := strings.Index(picUrl, "/uploadPic")
+	if pos == -1 {
+		return
+	}
 	finalPicUrl := config.PicSaverPath + picUrl[pos+10:]
 	err = os.Remove(finalPicUrl)
 	if err != nil {
